MO-584 unify API responses with one Offender object
