+++ app/controllers/prisoners_controller.rb
@@ -55,7 +55,7 @@ class PrisonersController < PrisonsApplicationController
 
   def image
     @prisoner = OffenderService.get_offender(params[:prisoner_id])
-    image_data = HmppsApi::PrisonApi::OffenderApi.get_image(@prisoner.booking_id)
+    image_data = @prisoner.get_image
 
     response.headers['Expires'] = 6.months.from_now.httpdate
     send_data image_data, type: 'image/jpg', disposition: 'inline'
