+++ spec/features/pom_caseload_feature_spec.rb
@@ -101,7 +101,7 @@ feature "view POM's caseload" do
       create(:responsibility, nomis_offender_id: o.fetch(:offenderNo), value: Responsibility::PROBATION)
     end
 
-    allow_any_instance_of(HmppsApi::OffenderBase).to receive(:handover_start_date).and_return(tomorrow)
+    allow_any_instance_of(HmppsApi::Offender).to receive(:handover_start_date).and_return(tomorrow)
 
     stub_user staff_id: nomis_staff_id
   end
